+++ packages/stargate/src/stargateclient.spec.ts
@@ -83,7 +83,7 @@ describe("StargateClient", () => {
   describe("connect", () => {
     it("works", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       expect(client).toBeTruthy();
       client.disconnect();
     });
@@ -92,14 +92,14 @@ describe("StargateClient", () => {
   describe("getChainId", () => {
     it("works", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       expect(await client.getChainId()).toEqual(simapp.chainId);
       client.disconnect();
     });
 
     it("caches chain ID", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const openedClient = client as unknown as PrivateStargateClient;
       const getCodeSpy = spyOn(openedClient.cometClient!, "status").and.callThrough();
 
@@ -115,7 +115,7 @@ describe("StargateClient", () => {
   describe("getHeight", () => {
     it("works", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const height1 = await client.getHeight();
       expect(height1).toBeGreaterThan(0);
@@ -131,7 +131,7 @@ describe("StargateClient", () => {
   describe("getAccount", () => {
     it("works for unused account", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const account = await client.getAccount(unused.address);
       assert(account);
@@ -147,7 +147,7 @@ describe("StargateClient", () => {
 
     it("works for account with pubkey and non-zero sequence", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const account = await client.getAccount(validator.delegatorAddress);
       assert(account);
@@ -163,7 +163,7 @@ describe("StargateClient", () => {
 
     it("returns null for non-existent address", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const account = await client.getAccount(nonExistentAddress);
       expect(account).toBeNull();
@@ -175,7 +175,7 @@ describe("StargateClient", () => {
   describe("getSequence", () => {
     it("works for unused account", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const account = await client.getSequence(unused.address);
       assert(account);
@@ -189,7 +189,7 @@ describe("StargateClient", () => {
 
     it("rejects for non-existent address", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       await expectAsync(client.getSequence(nonExistentAddress)).toBeRejectedWithError(
         /account '([a-z0-9]{10,90})' does not exist on chain/i,
@@ -202,7 +202,7 @@ describe("StargateClient", () => {
   describe("getBlock", () => {
     it("works for latest block", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const response = await client.getBlock();
 
       expect(response).toEqual(
@@ -226,7 +226,7 @@ describe("StargateClient", () => {
 
     it("works for block by height", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const height = (await client.getBlock()).header.height;
       const response = await client.getBlock(height - 1);
 
@@ -253,7 +253,7 @@ describe("StargateClient", () => {
   describe("getBalance", () => {
     it("works for different existing balances", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const response1 = await client.getBalance(unused.address, simapp.denomFee);
       expect(response1).toEqual({
@@ -271,7 +271,7 @@ describe("StargateClient", () => {
 
     it("returns 0 for non-existent balance", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const response = await client.getBalance(unused.address, "gintonic");
       expect(response).toEqual({
@@ -284,7 +284,7 @@ describe("StargateClient", () => {
 
     it("returns 0 for non-existent address", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const response = await client.getBalance(nonExistentAddress, simapp.denomFee);
       expect(response).toEqual({
@@ -299,7 +299,7 @@ describe("StargateClient", () => {
   describe("getAllBalances", () => {
     it("returns all balances for unused account", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const balances = await client.getAllBalances(unused.address);
       expect(balances).toEqual([
@@ -318,7 +318,7 @@ describe("StargateClient", () => {
 
     it("returns an empty list for non-existent account", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
 
       const balances = await client.getAllBalances(nonExistentAddress);
       expect(balances).toEqual([]);
@@ -330,7 +330,7 @@ describe("StargateClient", () => {
   describe("getBalanceStaked", () => {
     it("works", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const response = await client.getBalanceStaked(faucet.address0);
 
       expect(response).toEqual({ denom: "ustake", amount: "63474" });
@@ -342,7 +342,7 @@ describe("StargateClient", () => {
   describe("broadcastTx", () => {
     it("broadcasts a transaction", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const [{ address, pubkey: pubkeyBytes }] = await wallet.getAccounts();
       const pubkey = encodePubkey({
@@ -410,7 +410,7 @@ describe("StargateClient", () => {
 
     it("errors immediately for a CheckTx failure", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const [{ address, pubkey: pubkeyBytes }] = await wallet.getAccounts();
       const pubkey = encodePubkey({
@@ -491,7 +491,7 @@ describe("StargateClient", () => {
 
     it("respects user timeouts rather than RPC timeouts", async () => {
       pendingWithoutSlowSimapp();
-      const client = await StargateClient.connect(slowSimapp.tendermintUrl);
+      const client = await StargateClient.connect(slowSimapp.tendermintUrlHttp);
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const [{ address, pubkey: pubkeyBytes }] = await wallet.getAccounts();
       const pubkey = encodePubkey({
@@ -575,7 +575,7 @@ describe("StargateClient", () => {
   describe("broadcastTxSync", () => {
     it("broadcasts sync a transaction, to get transaction hash", async () => {
       pendingWithoutSimapp();
-      const client = await StargateClient.connect(simapp.tendermintUrl);
+      const client = await StargateClient.connect(simapp.tendermintUrlHttp);
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const [{ address, pubkey: pubkeyBytes }] = await wallet.getAccounts();
       const pubkey = encodePubkey({